Refactor enabling edge/proposed features
