+++ lib/ruby-next/language/edge.rb
@@ -1,0 +1,3 @@
+# frozen_string_literal: true
+
+# Load edge Ruby features
