+++ lib/ruby-next/version.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
 module RubyNext
-  VERSION = "0.5.3"
+  VERSION = "0.6.0-dev"
 end
