+++ spec/integration/fixtures/method_reference.rb
