--- spec/optional/language/fixtures/method_reference.rb
@@ -1,9 +1,0 @@
-module MethodReference
-  def self.positive?(num)
-    num > 0
-  end
-
-  def self.add(a, b)
-    a + b
-  end
-end
