--- spec/optional/language/method_reference_spec.rb
@@ -1,23 +1,0 @@
-# source: none
-
-require_relative 'fixtures/method_reference'
-
-describe "method reference .:" do
-  it "returns a Method" do
-    m = MethodReference.:add
-    m.class.should == Method
-  end
-
-  it "works" do
-    MethodReference.:add.call(2, 3).should == 5
-  end
-
-  it "works with to_proc" do
-    [-1, 2, -3].select(&MethodReference.:positive?).should == [2]
-  end
-
-  it "works with unary methods" do
-    m = 1.:-@
-    m.call.should == -1
-  end
-end
